+++ packages/graphql-persisted/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## Unreleased -->
 
+## 3.1.3 - 2022-02-02
+
+- No updates. Transitive dependency bump.
+
 ## 3.1.2 - 2022-02-01
 
 ### Changed